OPT: minor corrections on type issues
Former-commit-id: 23283ac19c3923d6e7b60efbd93b2f2a13364293
Former-commit-id: 68288d57bff3632cb65cf2e3a76c71c147491923

+++ scripts/sct_straighten_spinalcord.py
@@ -767,8 +767,8 @@ class SpinalCordStraightener(object):
                 # convert landmarks to INT
                 sct.printv('\nConvert landmarks to INT...', verbose)
                 from sct_convert import convert
-                convert('tmp.landmarks_straight.nii.gz', 'tmp.landmarks_straight.nii.gz', type='int8')
-                convert('tmp.landmarks_curved.nii.gz', 'tmp.landmarks_curved.nii.gz', type='int8')
+                convert('tmp.landmarks_straight.nii.gz', 'tmp.landmarks_straight.nii.gz', type='int32')
+                convert('tmp.landmarks_curved.nii.gz', 'tmp.landmarks_curved.nii.gz', type='int32')
 
                 # This stands to avoid overlapping between landmarks
                 # TODO: do symmetric removal
